Rename and test date functions
