+++ server/dates.ts
@@ -2,6 +2,16 @@ import { Arrays } from 'collection-fns';
 import { format, addDays, startOfWeek, parse, subMinutes } from 'date-fns';
 import { Config } from './app-config';
 
+export const getVisibleDates = (
+  config: Pick<Config, 'advanceBookingDays' | 'dataRetentionDays'>
+) => {
+  const today = new Date();
+  const start = addDays(today, -1 * config.dataRetentionDays);
+  return Arrays.init({ count: config.advanceBookingDays + config.dataRetentionDays }, (i) =>
+    format(addDays(start, i), 'yyyy-MM-dd')
+  );
+};
+
 export const getAvailableDates = (config: Pick<Config, 'advanceBookingDays'>) => {
   const today = new Date();
   return Arrays.init({ count: config.advanceBookingDays }, (i) =>
